Update readme, no Go, include plans for Deno
Really liking Deno and how it could be used for a wide range of
scripting. Still playing around with ideas on how to make it easier to
understand how all the pieces of chillbox fit together. Still have to
pretty much walk through the scripts to understand how it works.