Rename user data script, prep for initial ansible
The user-data script that is added to a newly deployed chillbox server.
It will install python and openssl and then output the encrypted
bootstrap file to the /root directory. Then an ansible playbook will
decrypt that file with the supplied bootstrap_chillbox_pass and start
off the chillbox init script.