Register letsencrypt account in terraform-010-infra
Only should register the letsencrypt account one time for the
environment. This also allows this registration to happen interactively.
The files generated via 'certbot register' are encrypted with the gpg
key in terraform-010-infra container and then decrypted to plaintext
when it is embedded in the bootstrap-chillbox-init-credentials.sh
script; which is encrypted with the symetric key.